Rename fakes to mocks and add docs
