+++ tests/integration/__init__.py
@@ -1,6 +1,11 @@
+"""
+The commander interacts with the file system (to be precise: the .mcmd dir in the user's home). To prevent these
+interactions during the integration tests, the modules need to be replaced with mocks, which is done in this package.
+"""
+
 import sys
 
-from tests.integration import fake_loader, fake_home
+from tests.integration import loader_mock, home_mock
 
-sys.modules['mcmd.config.loader'] = fake_loader
-sys.modules['mcmd.config.home'] = fake_home
+sys.modules['mcmd.config.loader'] = loader_mock
+sys.modules['mcmd.config.home'] = home_mock
